Query resultsdb for the results of the Atomic CI pipeline and display them
The way bodhi was querying resultsdb for test results was relying on the
specific structure of taskotron's results which lead to the results from
the Atomic CI pipeline to not be included.
With this commit we're querying for the specific structure of the Atomic
CI results so they are included in the "Automated tests" tab of the
update page.

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>